changed the realloc functions to prevent segfaulting

+++ libft/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: tblaase <tblaase@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/15 17:23:59 by tschmitt          #+#    #+#             */
-/*   Updated: 2021/11/30 22:51:02 by tblaase          ###   ########.fr       */
+/*   Updated: 2021/12/01 11:41:44 by tblaase          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,8 +108,8 @@ char	**ft_split_set(char *str, char *set);
 char	*ft_strset(char *str, char *set);
 char	**ft_str_arr_dup(char **arr);
 int		ft_str_arr_len(char **arr);
-char	*ft_realloc_str(char *str, size_t size);
-char	**ft_realloc_str_arr(char **arr, size_t size);
+char	*ft_realloc_str(char *str, int size);
+char	**ft_realloc_str_arr(char **arr, int size);
 char	*ft_strstrjoin(char *s1, char *s2, char *divider);
 int		ft_strclen(char *str, char c);
 char	*ft_append_len_div(char **dst, char *src, int len, char *div);
